# libs/api-client/src/generated/schemas.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export const RegisterDtoSchema = {
    type: 'object',
    properties: {
        email: {
            type: 'string',
            example: 'jane.doe@example.com'
        },
        password: {
            type: 'string',
            example: 'correct horse battery staple'
        },
        name: {
            type: 'string',
            example: 'Jane Doe'
        }
    },
    required: ['email', 'password', 'name']
} as const;

export const UserResponseDtoSchema = {
    type: 'object',
    properties: {
        id: {
            type: 'string'
        },
        email: {
            type: 'string'
        },
        name: {
            type: 'string'
        }
    },
    required: ['id', 'email', 'name']
} as const;

export const AuthResponseDtoSchema = {
    type: 'object',
    properties: {
        accessToken: {
            type: 'string'
        },
        refreshToken: {
            type: 'string'
        },
        tokenType: {
            type: 'string',
            example: 'Bearer'
        },
        expiresIn: {
            type: 'number',
            description: 'Access token lifetime in seconds'
        },
        user: {
            '$ref': '#/components/schemas/UserResponseDto'
        }
    },
    required: ['accessToken', 'refreshToken', 'tokenType', 'expiresIn', 'user']
} as const;

export const ErrorResponseDtoSchema = {
    type: 'object',
    properties: {
        statusCode: {
            type: 'number',
            example: 404
        },
        message: {
            type: 'string',
            example: 'Patient not found'
        }
    },
    required: ['statusCode', 'message']
} as const;

export const LoginDtoSchema = {
    type: 'object',
    properties: {
        email: {
            type: 'string',
            example: 'jane.doe@example.com'
        },
        password: {
            type: 'string',
            example: 'correct horse battery staple'
        }
    },
    required: ['email', 'password']
} as const;

export const RefreshRequestDtoSchema = {
    type: 'object',
    properties: {
        refreshToken: {
            type: 'string'
        }
    },
    required: ['refreshToken']
} as const;

export const LogoutRequestDtoSchema = {
    type: 'object',
    properties: {
        refreshToken: {
            type: 'string'
        }
    },
    required: ['refreshToken']
} as const;

export const PatientResponseDtoSchema = {
    type: 'object',
    properties: {
        id: {
            type: 'string'
        },
        mrn: {
            type: 'string'
        },
        firstName: {
            type: 'string'
        },
        lastName: {
            type: 'string'
        },
        dateOfBirth: {
            type: 'string',
            format: 'date',
            example: '1990-05-14'
        },
        gender: {
            type: 'string',
            enum: ['MALE', 'FEMALE', 'OTHER']
        },
        email: {
            type: 'string'
        },
        createdAt: {
            type: 'string',
            format: 'date-time'
        }
    },
    required: ['id', 'mrn', 'firstName', 'lastName', 'dateOfBirth', 'gender', 'email', 'createdAt']
} as const;

export const CreatePatientDtoSchema = {
    type: 'object',
    properties: {
        firstName: {
            type: 'string',
            example: 'Jane'
        },
        lastName: {
            type: 'string',
            example: 'Doe'
        },
        dateOfBirth: {
            type: 'string',
            example: '1990-05-14'
        },
        gender: {
            type: 'string',
            enum: ['MALE', 'FEMALE', 'OTHER']
        },
        email: {
            type: 'string',
            example: 'jane.doe@example.com'
        }
    },
    required: ['firstName', 'lastName', 'dateOfBirth', 'gender', 'email']
} as const;

export const UpdatePatientDtoSchema = {
    type: 'object',
    properties: {
        firstName: {
            type: 'string',
            example: 'Jane'
        },
        lastName: {
            type: 'string',
            example: 'Doe'
        },
        dateOfBirth: {
            type: 'string',
            example: '1990-05-14'
        },
        gender: {
            type: 'string',
            enum: ['MALE', 'FEMALE', 'OTHER']
        },
        email: {
            type: 'string',
            example: 'jane.doe@example.com'
        }
    }
} as const;